# src/routes/authRouter.js
const express = require("express");
const User = require("../models/user");
const bcrypt = require("bcrypt");
const { validateSignUpData } = require("../utils/validation");
const validator = require("validator");

const authRouter = express.Router();

authRouter.post("/signup", async (req, res) => {
  try {
    validateSignUpData(req);

    const {
      firstName,
      lastName,
      emailId,
      password,
      age,
      gender,
      photoUrl,
      about,
      skills,
    } = req.body;

    const hashedPassword = await bcrypt.hash(password, 10);

    const user = new User({
      firstName,
      lastName,
      emailId,
      password: hashedPassword,
      age,
      gender,
      photoUrl,
      about,
      skills,
    });

    const savedUser = await user.save();

    const token = await savedUser.getJWT();
    // res.cookie("token", token, {
    //   expires: new Date(Date.now() + 8 * 3600000),
    // });
    res.cookie("token", token, {
      httpOnly: true,
      secure: true,
      sameSite: "None",
      expires: new Date(Date.now() + 8 * 3600000),
    });

    res
      .status(201)
      .send({ message: "User registered successfully!", user: savedUser });
  } catch (error) {
    res
      .status(400)
      .send({ message: "Error registering the user", ERROR: error.message });
  }
});

authRouter.post("/login", async (req, res) => {
  try {
    const { emailId, password } = req.body;

    if (!validator.isEmail(emailId)) {
      throw new Error("Invalid email address");
    }

    const user = await User.findOne({ emailId });

    if (!user) {
      throw new Error("Invalid credentials");
    }

    const isPasswordValid = await user.validatePassword(password);

    if (isPasswordValid) {
      const token = await user.getJWT();
      // res.cookie("token", token, {
      //   expires: new Date(Date.now() + 8 * 3600000),
      // });
      res.cookie("token", token, {
        httpOnly: true,
        secure: true,
        sameSite: "None",
        expires: new Date(Date.now() + 8 * 3600000),
      });
      res.status(200).json({ message: "User logged in successfully!", user });
    } else {
      throw new Error("Invalid credentials");
    }
  } catch (error) {
    res
      .status(400)
      .send({ message: "Error logging in the user", ERROR: error.message });
  }
});

authRouter.post("/logout", async (req, res) => {
  try {
    res.cookie("token", null, { expires: new Date(Date.now()) });
    res.status(200).send("User logged out successfully!");
  } catch (error) {
    res
      .status(400)
      .send({ message: "Error logging in the user", ERROR: error.message });
  }
});

module.exports = authRouter;