# swagger.yaml
openapi: 3.0.0
info:
  title: CodeSwipe API
  version: 1.0.0
  description: |

    API documentation for the CodeSwipe backend application.

    CodeSwipe is a developer-focused social networking platform that enables users to connect, collaborate, and grow their professional network. 

    This backend API supports a wide range of features including user authentication, profile management, connection requests, real-time chat, and membership upgrades via Razorpay-integrated payment processing.

    The platform offers tiered membership options—Gold and Silver—with different levels of access and privileges. 
    Scheduled background jobs (via cron) are used for periodic tasks such as data maintenance and notification handling.

    This documentation outlines all public endpoints used by the client application to interact with the CodeSwipe backend.
servers:
  - url: https://codeswipe-backend.onrender.com
    description: Server
paths:
  /signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - emailId
                - password
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                emailId:
                  type: string
                password:
                  type: string
                  format: password
                age:
                  type: integer
                gender:
                  type: string
                photoUrl:
                  type: string
                about:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Error registering the user
  /login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emailId
                - password
              properties:
                emailId:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User logged in successfully
        '400':
          description: Error logging in the user
  /logout:
    post:
      tags:
        - Authentication
      summary: Log out the current user
      responses:
        '200':
          description: User logged out successfully
        '400':
          description: Error logging out the user
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token